清除第三方库，使用原生的 SoundPool 解析音频

+++ src/com/gracecode/android/rain/receiver/PlayBroadcastReceiver.java
@@ -92,7 +92,7 @@ public abstract class PlayBroadcastReceiver extends BroadcastReceiver {
 
     abstract public void onStop();
 
-    abstract public void onSetVolume(int track, int volume);
+    abstract public void onSetVolume(int track, float volume);
 
     abstract public void onSetPresets(float[] presets);
